convert headers to reflect consistent mem access
address, size, value

	modified:   include/csx_data.h
	modified:   include/csx_mem.h
	modified:   include/csx_mmio.h
	modified:   include/csx_mmio_reg.h
	modified:   include/csx_mmio_reg_t.h
	modified:   include/csx_test_utility.h
	modified:   include/soc.h
	modified:   include/soc_core_arm_decode.h
	modified:   include/soc_core_utility.h
	modified:   include/soc_mmio.h
	modified:   include/soc_mmio_trace.h
	modified:   include/soc_mmu.h
	modified:   include/soc_nnd_flash.h

+++ include/csx_test_utility.h
@@ -11,7 +11,7 @@ void _assert_cpsr_xpsr(csx_test_p t, uint cpsr, uint xpsr);
 void _assert_cpsr_xpsr_mask(csx_test_p t, uint cpsr, uint xpsr, uint mask);
 void _assert_nzc(csx_test_p t, int n, int z, int c);
 void _assert_nzcv(csx_test_p t, int n, int z, int c, int v);
-void _cxx(csx_test_p t, uint32_t value, uint8_t size);
+void _cxx(csx_test_p t, uint32_t value, size_t size);
 uint32_t _test_value(uint8_t i);
 
 /* **** */
